requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
twilio==8.10.0
python-dotenv==1.0.0